requirements.txt
# Orchestration
langgraph>=0.2

# LLM
groq>=0.9

# Embeddings (local, no API calls)
sentence-transformers>=3.0

# Vector search (local file)
faiss-cpu>=1.8

# HTTP (Reddit public JSON API — no auth needed)
requests>=2.31

# PDF extraction for quality dataset seeding
pypdf>=4.0

# Optional browser collector for Facebook mode=selenium
selenium>=4.0

# Language detection
langdetect>=1.0.9

# Retries
tenacity>=8.2

# Validation
pydantic>=2.0

# Config
pyyaml>=6.0
python-dotenv>=1.0

# Timezone
pytz>=2024.1

# Database
sqlalchemy>=2.0
alembic>=1.13

# Backend
fastapi>=0.111
uvicorn[standard]>=0.30
sse-starlette>=2.1

# Testing
pytest>=8.0
